Split up /authorize request generation (EAR Pre-Req #2) (#7623)
Splits up /authorize request generation into separate functions for 
1. Standard params (included on all requests)
2. Browser auth code params 
3. Node auth code params

Note: #7633 adds another for Browser EAR
Note2: Future (uncommitted) work to move /token and /logout to a similar
pattern for extensibility and bundle size improvements
