+++ change/@azure-msal-browser-81b8d13b-0d90-4a64-bf38-09f56a789bd2.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Refactor /authorize request generation",
+  "packageName": "@azure/msal-browser",
+  "email": "thomas.norling@microsoft.com",
+  "dependentChangeType": "patch"
+}
