+++ change/@azure-msal-common-05af4dd2-f51f-46cd-a212-5e0e5b64288b.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Refactor /authorize request generation",
+  "packageName": "@azure/msal-common",
+  "email": "thomas.norling@microsoft.com",
+  "dependentChangeType": "patch"
+}
