+++ change/@azure-msal-node-c5c3efce-d2b0-47b3-ac24-867539845fb6.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Refactor /authorize request generation",
+  "packageName": "@azure/msal-node",
+  "email": "thomas.norling@microsoft.com",
+  "dependentChangeType": "patch"
+}
