+++ lib/msal-browser/test/interaction_client/SilentAuthCodeClient.spec.ts
@@ -34,6 +34,7 @@ import {
 } from "../../src/index.js";
 import { InteractionHandler } from "../../src/interaction_handler/InteractionHandler.js";
 import { FetchClient } from "../../src/network/FetchClient.js";
+import * as AuthorizeProtocol from "../../src/protocol/Authorize.js";
 import { TestTimeUtils } from "msal-test-utils";
 
 describe("SilentAuthCodeClient", () => {
@@ -137,8 +138,8 @@ describe("SilentAuthCodeClient", () => {
                 tokenType: AuthenticationScheme.BEARER,
             };
             jest.spyOn(
-                AuthorizationCodeClient.prototype,
-                "getAuthCodeUrl"
+                AuthorizeProtocol,
+                "getAuthCodeRequestUrl"
             ).mockResolvedValue(testNavUrl);
             const handleCodeSpy = jest
                 .spyOn(
