+++ lib/msal-common/src/response/ResponseHandler.ts
@@ -9,7 +9,6 @@ import {
     ClientAuthErrorCodes,
     createClientAuthError,
 } from "../error/ClientAuthError.js";
-import { ServerAuthorizationCodeResponse } from "./ServerAuthorizationCodeResponse.js";
 import { Logger } from "../logger/Logger.js";
 import { ServerError } from "../error/ServerError.js";
 import { ScopeSet } from "../request/ScopeSet.js";
@@ -54,19 +53,6 @@ import {
 import * as CacheHelpers from "../cache/utils/CacheHelpers.js";
 import * as TimeUtils from "../utils/TimeUtils.js";
 
-function parseServerErrorNo(
-    serverResponse: ServerAuthorizationCodeResponse
-): string | undefined {
-    const errorCodePrefix = "code=";
-    const errorCodePrefixIndex =
-        serverResponse.error_uri?.lastIndexOf(errorCodePrefix);
-    return errorCodePrefixIndex && errorCodePrefixIndex >= 0
-        ? serverResponse.error_uri?.substring(
-              errorCodePrefixIndex + errorCodePrefix.length
-          )
-        : undefined;
-}
-
 /**
  * Class that handles response parsing.
  * @internal
@@ -97,90 +83,6 @@ export class ResponseHandler {
         this.serializableCache = serializableCache;
         this.persistencePlugin = persistencePlugin;
         this.performanceClient = performanceClient;
-    }
-
-    /**
-     * Function which validates server authorization code response.
-     * @param serverResponseHash
-     * @param requestState
-     * @param cryptoObj
-     */
-    validateServerAuthorizationCodeResponse(
-        serverResponse: ServerAuthorizationCodeResponse,
-        requestState: string
-    ): void {
-        if (!serverResponse.state || !requestState) {
-            throw serverResponse.state
-                ? createClientAuthError(
-                      ClientAuthErrorCodes.stateNotFound,
-                      "Cached State"
-                  )
-                : createClientAuthError(
-                      ClientAuthErrorCodes.stateNotFound,
-                      "Server State"
-                  );
-        }
-
-        let decodedServerResponseState: string;
-        let decodedRequestState: string;
-
-        try {
-            decodedServerResponseState = decodeURIComponent(
-                serverResponse.state
-            );
-        } catch (e) {
-            throw createClientAuthError(
-                ClientAuthErrorCodes.invalidState,
-                serverResponse.state
-            );
-        }
-
-        try {
-            decodedRequestState = decodeURIComponent(requestState);
-        } catch (e) {
-            throw createClientAuthError(
-                ClientAuthErrorCodes.invalidState,
-                serverResponse.state
-            );
-        }
-
-        if (decodedServerResponseState !== decodedRequestState) {
-            throw createClientAuthError(ClientAuthErrorCodes.stateMismatch);
-        }
-
-        // Check for error
-        if (
-            serverResponse.error ||
-            serverResponse.error_description ||
-            serverResponse.suberror
-        ) {
-            const serverErrorNo = parseServerErrorNo(serverResponse);
-            if (
-                isInteractionRequiredError(
-                    serverResponse.error,
-                    serverResponse.error_description,
-                    serverResponse.suberror
-                )
-            ) {
-                throw new InteractionRequiredAuthError(
-                    serverResponse.error || "",
-                    serverResponse.error_description,
-                    serverResponse.suberror,
-                    serverResponse.timestamp || "",
-                    serverResponse.trace_id || "",
-                    serverResponse.correlation_id || "",
-                    serverResponse.claims || "",
-                    serverErrorNo
-                );
-            }
-
-            throw new ServerError(
-                serverResponse.error || "",
-                serverResponse.error_description,
-                serverResponse.suberror,
-                serverErrorNo
-            );
-        }
     }
 
     /**
