+++ lib/msal-common/src/utils/UrlUtils.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { ServerAuthorizationCodeResponse } from "../response/ServerAuthorizationCodeResponse.js";
+import { AuthorizeResponse } from "../response/AuthorizeResponse.js";
 import {
     ClientAuthErrorCodes,
     createClientAuthError,
@@ -31,7 +31,7 @@ export function stripLeadingHashOrQuery(responseString: string): string {
  */
 export function getDeserializedResponse(
     responseString: string
-): ServerAuthorizationCodeResponse | null {
+): AuthorizeResponse | null {
     // Check if given hash is empty
     if (!responseString || responseString.indexOf("=") < 0) {
         return null;
@@ -40,8 +40,9 @@ export function getDeserializedResponse(
         // Strip the # or ? symbol if present
         const normalizedResponse = stripLeadingHashOrQuery(responseString);
         // If # symbol was not present, above will return empty string, so give original hash value
-        const deserializedHash: ServerAuthorizationCodeResponse =
-            Object.fromEntries(new URLSearchParams(normalizedResponse));
+        const deserializedHash: AuthorizeResponse = Object.fromEntries(
+            new URLSearchParams(normalizedResponse)
+        );
 
         // Check for known response properties
         if (
