+++ lib/msal-common/test/client/AuthorizationCodeClient.spec.ts
@@ -1,5 +1,4 @@
 import {
-    ALTERNATE_OPENID_CONFIG_RESPONSE,
     AUTHENTICATION_RESULT,
     DEFAULT_OPENID_CONFIG_RESPONSE,
     TEST_CONFIG,
@@ -7,10 +6,8 @@ import {
     TEST_URIS,
     TEST_DATA_CLIENT_INFO,
     RANDOM_TEST_GUID,
-    TEST_STATE_VALUES,
     TEST_POP_VALUES,
     POP_AUTHENTICATION_RESULT,
-    TEST_ACCOUNT_INFO,
     CORS_SIMPLE_REQUEST_HEADERS,
     TEST_SSH_VALUES,
     AUTHENTICATION_RESULT_WITH_HEADERS,
@@ -19,8 +16,6 @@ import {
 import { ClientConfiguration } from "../../src/config/ClientConfiguration.js";
 import { BaseClient } from "../../src/client/BaseClient.js";
 import {
-    PromptValue,
-    ResponseMode,
     AuthenticationScheme,
     ThrottlingConstants,
     Constants,
@@ -28,13 +23,11 @@ import {
     ONE_DAY_IN_MS,
 } from "../../src/utils/Constants.js";
 import * as AADServerParamKeys from "../../src/constants/AADServerParamKeys.js";
-import { ClientTestUtils, MockStorageClass } from "./ClientTestUtils.js";
+import { ClientTestUtils } from "./ClientTestUtils.js";
 import { TestError } from "../test_kit/TestErrors.js";
 import { Authority } from "../../src/authority/Authority.js";
 import { AuthorizationCodeClient } from "../../src/client/AuthorizationCodeClient.js";
-import { CommonAuthorizationUrlRequest } from "../../src/request/CommonAuthorizationUrlRequest.js";
 import { TokenClaims } from "../../src/account/TokenClaims.js";
-import { ServerError } from "../../src/error/ServerError.js";
 import { CommonAuthorizationCodeRequest } from "../../src/request/CommonAuthorizationCodeRequest.js";
 import * as AuthToken from "../../src/account/AuthToken.js";
 import {
@@ -42,13 +35,11 @@ import {
     createClientAuthError,
 } from "../../src/error/ClientAuthError.js";
 import {
-    AuthError,
     CcsCredentialType,
     ClientConfigurationErrorCodes,
     createClientConfigurationError,
 } from "../../src/index.js";
 import { ProtocolMode } from "../../src/authority/ProtocolMode.js";
-import { MockPerformanceClient } from "../telemetry/PerformanceClient.spec.js";
 
 describe("AuthorizationCodeClient unit tests", () => {
     afterEach(() => {
@@ -67,1388 +58,6 @@ describe("AuthorizationCodeClient unit tests", () => {
             expect(client).not.toBeNull();
             expect(client instanceof AuthorizationCodeClient).toBe(true);
             expect(client instanceof BaseClient).toBe(true);
-        });
-    });
-
-    describe("Authorization url creation", () => {
-        it("Creates an authorization url with default parameters", async () => {
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.QUERY,
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: TEST_CONFIG.DEFAULT_SCOPES,
-                codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
-                codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(Constants.DEFAULT_AUTHORITY)).toBe(true);
-            expect(
-                loginUrl.includes(
-                    DEFAULT_OPENID_CONFIG_RESPONSE.body.authorization_endpoint.replace(
-                        "{tenant}",
-                        "common"
-                    )
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SCOPE}=${Constants.OPENID_SCOPE}%20${Constants.PROFILE_SCOPE}%20${Constants.OFFLINE_ACCESS_SCOPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_TYPE}=${Constants.CODE_RESPONSE_TYPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CLIENT_ID}=${TEST_CONFIG.MSAL_CLIENT_ID}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.REDIRECT_URI}=${encodeURIComponent(
-                        TEST_URIS.TEST_REDIRECT_URI_LOCALHOST
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_MODE}=${encodeURIComponent(
-                        ResponseMode.QUERY
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CODE_CHALLENGE}=${encodeURIComponent(
-                        TEST_CONFIG.TEST_CHALLENGE
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${
-                        AADServerParamKeys.CODE_CHALLENGE_METHOD
-                    }=${encodeURIComponent(
-                        Constants.S256_CODE_CHALLENGE_METHOD
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.X_APP_NAME}=${TEST_CONFIG.applicationName}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.X_APP_VER}=${TEST_CONFIG.applicationVersion}`
-                )
-            ).toBe(true);
-        });
-
-        it("Creates an authorization url passing in optional parameters", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FORM_POST,
-                codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
-                codeChallengeMethod: TEST_CONFIG.CODE_CHALLENGE_METHOD,
-                state: TEST_CONFIG.STATE,
-                prompt: PromptValue.LOGIN,
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                domainHint: TEST_CONFIG.DOMAIN_HINT,
-                claims: TEST_CONFIG.CLAIMS,
-                nonce: TEST_CONFIG.NONCE,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(TEST_CONFIG.validAuthority)).toBe(true);
-            expect(
-                loginUrl.includes(
-                    DEFAULT_OPENID_CONFIG_RESPONSE.body.authorization_endpoint.replace(
-                        "{tenant}",
-                        "common"
-                    )
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SCOPE}=${TEST_CONFIG.DEFAULT_GRAPH_SCOPE}%20${Constants.OPENID_SCOPE}%20${Constants.PROFILE_SCOPE}%20${Constants.OFFLINE_ACCESS_SCOPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_TYPE}=${Constants.CODE_RESPONSE_TYPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CLIENT_ID}=${TEST_CONFIG.MSAL_CLIENT_ID}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.REDIRECT_URI}=${encodeURIComponent(
-                        TEST_URIS.TEST_REDIRECT_URI_LOCALHOST
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_MODE}=${encodeURIComponent(
-                        ResponseMode.FORM_POST
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.STATE}=${encodeURIComponent(
-                        TEST_CONFIG.STATE
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.PROMPT}=${PromptValue.LOGIN}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.NONCE}=${encodeURIComponent(
-                        TEST_CONFIG.NONCE
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CODE_CHALLENGE}=${encodeURIComponent(
-                        TEST_CONFIG.TEST_CHALLENGE
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${
-                        AADServerParamKeys.CODE_CHALLENGE_METHOD
-                    }=${encodeURIComponent(TEST_CONFIG.CODE_CHALLENGE_METHOD)}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.DOMAIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.DOMAIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CLAIMS}=${encodeURIComponent(
-                        TEST_CONFIG.CLAIMS
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.X_APP_NAME}=${TEST_CONFIG.applicationName}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.X_APP_VER}=${TEST_CONFIG.applicationVersion}`
-                )
-            ).toBe(true);
-        });
-
-        it("Adds CCS entry if loginHint is provided", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const mockPerfClient = new MockPerformanceClient();
-            const client = new AuthorizationCodeClient(config, mockPerfClient);
-            let resEvents;
-            mockPerfClient.addPerformanceCallback((events) => {
-                resEvents = events;
-            });
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                prompt: PromptValue.LOGIN,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const rootMeasurement = mockPerfClient.startMeasurement(
-                "root-measurement",
-                authCodeUrlRequest.correlationId
-            );
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${HeaderNames.CCS_HEADER}=${encodeURIComponent(
-                        `UPN:${TEST_CONFIG.LOGIN_HINT}`
-                    )}`
-                )
-            ).toBe(true);
-            rootMeasurement.end({ success: true });
-            // @ts-ignore
-            const event = resEvents[0];
-            expect(event.loginHintFromRequest).toBeTruthy();
-            expect(event.loginHintFromUpn).toBeFalsy();
-            expect(event.loginHintFromClaim).toBeFalsy();
-        });
-
-        it("Adds CCS entry if account is provided", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-            const testAccount = TEST_ACCOUNT_INFO;
-            // @ts-ignore
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                    | "login_hint"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                prompt: PromptValue.NONE,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        testTokenClaims.sid
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${HeaderNames.CCS_HEADER}=${encodeURIComponent(
-                        `Oid:${TEST_DATA_CLIENT_INFO.TEST_UID}@${TEST_DATA_CLIENT_INFO.TEST_UTID}`
-                    )}`
-                )
-            ).toBe(true);
-        });
-
-        it("prefers login_hint claim over sid/upn if both provided", async () => {
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const mockPerfClient = new MockPerformanceClient();
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config, mockPerfClient);
-            let resEvents;
-            mockPerfClient.addPerformanceCallback((events) => {
-                resEvents = events;
-            });
-            const testAccount = TEST_ACCOUNT_INFO;
-            // @ts-ignore
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-                login_hint: "opaque-login-hint-claim",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                prompt: PromptValue.NONE,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const rootMeasurement = mockPerfClient.startMeasurement(
-                "root-measurement",
-                authCodeUrlRequest.correlationId
-            );
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        testTokenClaims.sid
-                    )}`
-                )
-            ).toBe(false);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        testTokenClaims.login_hint
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${HeaderNames.CCS_HEADER}=${encodeURIComponent(
-                        `Oid:${TEST_DATA_CLIENT_INFO.TEST_UID}@${TEST_DATA_CLIENT_INFO.TEST_UTID}`
-                    )}`
-                )
-            ).toBe(true);
-
-            rootMeasurement.end({ success: true });
-            // @ts-ignore
-            const event = resEvents[0];
-            expect(event.loginHintFromUpn).toBeFalsy();
-            expect(event.loginHintFromClaim).toBeTruthy();
-            expect(event.loginHintFromRequest).toBeFalsy();
-            expect(event.domainHintFromRequest).toBeFalsy();
-            expect(event.sidFromClaim).toBeFalsy();
-            expect(event.sidFromRequest).toBeFalsy();
-        });
-
-        it("skips login_hint claim if domainHint param is set", async () => {
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const mockPerfClient = new MockPerformanceClient();
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config, mockPerfClient);
-            let resEvents;
-            mockPerfClient.addPerformanceCallback((events) => {
-                resEvents = events;
-            });
-            const testAccount = TEST_ACCOUNT_INFO;
-            // @ts-ignore
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-                login_hint: "opaque-login-hint-claim",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-                domainHint: TEST_CONFIG.DOMAIN_HINT,
-            };
-            const rootMeasurement = mockPerfClient.startMeasurement(
-                "root-measurement",
-                authCodeUrlRequest.correlationId
-            );
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        testTokenClaims.sid
-                    )}`
-                )
-            ).toBe(false);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        testAccount.username
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.DOMAIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.DOMAIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${HeaderNames.CCS_HEADER}=${encodeURIComponent(
-                        `Oid:${TEST_DATA_CLIENT_INFO.TEST_UID}@${TEST_DATA_CLIENT_INFO.TEST_UTID}`
-                    )}`
-                )
-            ).toBe(true);
-
-            rootMeasurement.end({ success: true });
-            // @ts-ignore
-            const event = resEvents[0];
-            expect(event.loginHintFromUpn).toBeTruthy();
-            expect(event.loginHintFromClaim).toBeFalsy();
-            expect(event.loginHintFromRequest).toBeFalsy();
-            expect(event.domainHintFromRequest).toBeTruthy();
-            expect(event.sidFromClaim).toBeFalsy();
-            expect(event.sidFromRequest).toBeFalsy();
-        });
-
-        it("picks up both loginHint and domainHint params", async () => {
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-            const testAccount = TEST_ACCOUNT_INFO;
-            // @ts-ignore
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-                login_hint: "opaque-login-hint-claim",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-                domainHint: TEST_CONFIG.DOMAIN_HINT,
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        testTokenClaims.sid
-                    )}`
-                )
-            ).toBe(false);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.DOMAIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.DOMAIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-        });
-
-        it("Prefers sid over loginHint if both provided and prompt=None", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const mockPerfClient = new MockPerformanceClient();
-            const client = new AuthorizationCodeClient(config, mockPerfClient);
-            let resEvents;
-            mockPerfClient.addPerformanceCallback((events) => {
-                resEvents = events;
-            });
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                prompt: PromptValue.NONE,
-                sid: TEST_CONFIG.SID,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const rootMeasurement = mockPerfClient.startMeasurement(
-                "root-measurement",
-                authCodeUrlRequest.correlationId
-            );
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl).toEqual(
-                expect.not.arrayContaining([
-                    `${AADServerParamKeys.LOGIN_HINT}=`,
-                ])
-            );
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        TEST_CONFIG.SID
-                    )}`
-                )
-            ).toBe(true);
-
-            rootMeasurement.end({ success: true });
-            // @ts-ignore
-            const event = resEvents[0];
-            expect(event.loginHintFromRequest).toBeFalsy();
-            expect(event.loginHintFromClaim).toBeFalsy();
-            expect(event.loginHintFromUpn).toBeFalsy();
-            expect(event.domainHintFromRequest).toBeFalsy();
-            expect(event.sidFromRequest).toBeTruthy();
-            expect(event.sidFromRequest).toBeTruthy();
-            expect(event.prompt).toEqual(PromptValue.NONE);
-        });
-
-        it("Prefers loginHint over sid if both provided and prompt!=None", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                prompt: PromptValue.LOGIN,
-                sid: TEST_CONFIG.SID,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Ignores sid if prompt!=None", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const mockPerfClient = new MockPerformanceClient();
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config, mockPerfClient);
-            let resEvents;
-            mockPerfClient.addPerformanceCallback((events) => {
-                resEvents = events;
-            });
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                prompt: PromptValue.LOGIN,
-                sid: TEST_CONFIG.SID,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const rootMeasurement = mockPerfClient.startMeasurement(
-                "root-measurement",
-                authCodeUrlRequest.correlationId
-            );
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(`${AADServerParamKeys.LOGIN_HINT}=`)).toBe(
-                false
-            );
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-
-            rootMeasurement.end({ success: true });
-            // @ts-ignore
-            const event = resEvents[0];
-            expect(event.loginHintFromUpn).toBeFalsy();
-            expect(event.loginHintFromClaim).toBeFalsy();
-            expect(event.loginHintFromRequest).toBeFalsy();
-            expect(event.domainHintFromRequest).toBeFalsy();
-            expect(event.sidFromClaim).toBeFalsy();
-            expect(event.sidFromRequest).toBeFalsy();
-            expect(event.prompt).toEqual(PromptValue.LOGIN);
-        });
-
-        it("Prefers loginHint over Account if both provided and account does not have token claims", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                account: TEST_ACCOUNT_INFO,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_ACCOUNT_INFO.username
-                    )}`
-                )
-            ).toBe(false);
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Uses sid from account if not provided in request and prompt=None, overrides login_hint", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-            const testAccount = TEST_ACCOUNT_INFO;
-            // @ts-ignore
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                prompt: PromptValue.NONE,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SID}=${encodeURIComponent(
-                        testTokenClaims.sid
-                    )}`
-                )
-            ).toBe(true);
-            expect(loginUrl.includes(`${AADServerParamKeys.LOGIN_HINT}=`)).toBe(
-                false
-            );
-        });
-
-        it("Uses loginHint instead of sid from account prompt!=None", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-            const testAccount = TEST_ACCOUNT_INFO;
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                    | "iat"
-                    | "x5c_ca"
-                    | "ts"
-                    | "at"
-                    | "u"
-                    | "p"
-                    | "m"
-                    | "login_hint"
-                    | "aud"
-                    | "nbf"
-                    | "roles"
-                    | "amr"
-                    | "idp"
-                    | "auth_time"
-                    | "tfp"
-                    | "acr"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                sid: "testSid",
-                tenant_region_scope: "test_tenant_region_scope",
-                tenant_region_sub_scope: "test_tenant_region_sub_scope",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                prompt: PromptValue.LOGIN,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-        });
-
-        it("Uses login_hint instead of username if sid is not present in token claims for account or request", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-            const testAccount = TEST_ACCOUNT_INFO;
-            const testTokenClaims: Required<
-                Omit<
-                    TokenClaims,
-                    | "home_oid"
-                    | "upn"
-                    | "cloud_instance_host_name"
-                    | "cnf"
-                    | "emails"
-                    | "sid"
-                    | "iat"
-                    | "x5c_ca"
-                    | "ts"
-                    | "at"
-                    | "u"
-                    | "p"
-                    | "m"
-                    | "login_hint"
-                    | "aud"
-                    | "nbf"
-                    | "roles"
-                    | "amr"
-                    | "idp"
-                    | "auth_time"
-                    | "tfp"
-                    | "acr"
-                >
-            > = {
-                ver: "2.0",
-                iss: `${TEST_URIS.DEFAULT_INSTANCE}9188040d-6c67-4c5b-b112-36a304b66dad/v2.0`,
-                sub: "AAAAAAAAAAAAAAAAAAAAAIkzqFVrSaSaFHy782bbtaQ",
-                exp: 1536361411,
-                name: "Abe Lincoln",
-                preferred_username: "AbeLi@microsoft.com",
-                oid: "00000000-0000-0000-66f3-3332eca7ea81",
-                tid: "3338040d-6c67-4c5b-b112-36a304b66dad",
-                nonce: "123523",
-                tenant_region_scope: "test_tenant_region_scope",
-                tenant_region_sub_scope: "test_tenant_region_sub_scope",
-            };
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: {
-                    ...testAccount,
-                    idTokenClaims: testTokenClaims,
-                },
-                loginHint: TEST_CONFIG.LOGIN_HINT,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_CONFIG.LOGIN_HINT
-                    )}`
-                )
-            ).toBe(true);
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Sets login_hint to Account.username if login_hint and sid are not provided", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: TEST_ACCOUNT_INFO,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.LOGIN_HINT}=${encodeURIComponent(
-                        TEST_ACCOUNT_INFO.username
-                    )}`
-                )
-            ).toBe(true);
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Ignores Account if prompt is select_account", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                account: TEST_ACCOUNT_INFO,
-                prompt: "select_account",
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(`${AADServerParamKeys.LOGIN_HINT}=`)).toBe(
-                false
-            );
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Ignores loginHint if prompt is select_account", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                loginHint: "testaccount@microsoft.com",
-                prompt: "select_account",
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(`${AADServerParamKeys.LOGIN_HINT}=`)).toBe(
-                false
-            );
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Ignores sid if prompt is select_account", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: [
-                    ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                    ...TEST_CONFIG.DEFAULT_SCOPES,
-                ],
-                sid: "testsid",
-                prompt: "select_account",
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.includes(`${AADServerParamKeys.LOGIN_HINT}=`)).toBe(
-                false
-            );
-            expect(loginUrl.includes(`${AADServerParamKeys.SID}=`)).toBe(false);
-        });
-
-        it("Creates a login URL with scopes from given token request", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(ALTERNATE_OPENID_CONFIG_RESPONSE.body);
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const testScope1 = "testscope1";
-            const testScope2 = "testscope2";
-            const loginRequest: CommonAuthorizationUrlRequest = {
-                redirectUri: TEST_URIS.TEST_REDIR_URI,
-                scopes: [testScope1, testScope2],
-                codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
-                codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.FRAGMENT,
-            };
-
-            const loginUrl = await client.getAuthCodeUrl(loginRequest);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SCOPE}=${encodeURIComponent(
-                        `${testScope1} ${testScope2}`
-                    )}`
-                )
-            ).toBe(true);
-        });
-
-        it("Does not append an extra '?' when the authorization endpoint already contains a query string", async () => {
-            // Override with alternate authority openid_config
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromHardcodedValues"
-            ).mockReturnValue({
-                token_endpoint:
-                    "https://login.windows.net/common/oauth2/v2.0/token?param1=value1",
-                issuer: "https://login.windows.net/{tenantid}/v2.0",
-                userinfo_endpoint: "https://graph.microsoft.com/oidc/userinfo",
-                authorization_endpoint:
-                    "https://login.windows.net/common/oauth2/v2.0/authorize?param1=value1",
-                end_session_endpoint:
-                    "https://login.windows.net/common/oauth2/v2.0/logout?param1=value1",
-            });
-
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-                authority: TEST_CONFIG.validAuthority,
-                responseMode: ResponseMode.QUERY,
-                redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-                scopes: TEST_CONFIG.DEFAULT_SCOPES,
-                codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
-                codeChallengeMethod: Constants.S256_CODE_CHALLENGE_METHOD,
-                correlationId: RANDOM_TEST_GUID,
-                authenticationScheme: AuthenticationScheme.BEARER,
-            };
-            const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-            expect(loginUrl.split("?").length).toEqual(2);
-            expect(loginUrl.includes(`param1=value1`)).toBe(true);
-            expect(
-                loginUrl.includes(
-                    ALTERNATE_OPENID_CONFIG_RESPONSE.body.authorization_endpoint.replace(
-                        "{tenant}",
-                        "common"
-                    )
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.SCOPE}=${Constants.OPENID_SCOPE}%20${Constants.PROFILE_SCOPE}%20${Constants.OFFLINE_ACCESS_SCOPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_TYPE}=${Constants.CODE_RESPONSE_TYPE}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CLIENT_ID}=${TEST_CONFIG.MSAL_CLIENT_ID}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.REDIRECT_URI}=${encodeURIComponent(
-                        TEST_URIS.TEST_REDIRECT_URI_LOCALHOST
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.RESPONSE_MODE}=${encodeURIComponent(
-                        ResponseMode.QUERY
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${AADServerParamKeys.CODE_CHALLENGE}=${encodeURIComponent(
-                        TEST_CONFIG.TEST_CHALLENGE
-                    )}`
-                )
-            ).toBe(true);
-            expect(
-                loginUrl.includes(
-                    `${
-                        AADServerParamKeys.CODE_CHALLENGE_METHOD
-                    }=${encodeURIComponent(
-                        Constants.S256_CODE_CHALLENGE_METHOD
-                    )}`
-                )
-            ).toBe(true);
-        });
-    });
-
-    it("Adds req-cnf as needed", async () => {
-        // Override with alternate authority openid_config
-        jest.spyOn(
-            Authority.prototype,
-            <any>"getEndpointMetadataFromNetwork"
-        ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
-
-        const config: ClientConfiguration =
-            await ClientTestUtils.createTestClientConfiguration();
-        const client = new AuthorizationCodeClient(config);
-
-        if (!config.cryptoInterface) {
-            throw TestError.createTestSetupError(
-                "configuration cryptoInterface not initialized correctly."
-            );
-        }
-
-        const authCodeUrlRequest: CommonAuthorizationUrlRequest = {
-            redirectUri: TEST_URIS.TEST_REDIRECT_URI_LOCALHOST,
-            scopes: [
-                ...TEST_CONFIG.DEFAULT_GRAPH_SCOPE,
-                ...TEST_CONFIG.DEFAULT_SCOPES,
-            ],
-            authority: TEST_CONFIG.validAuthority,
-            responseMode: ResponseMode.FORM_POST,
-            codeChallenge: TEST_CONFIG.TEST_CHALLENGE,
-            codeChallengeMethod: TEST_CONFIG.CODE_CHALLENGE_METHOD,
-            state: TEST_CONFIG.STATE,
-            prompt: PromptValue.LOGIN,
-            loginHint: TEST_CONFIG.LOGIN_HINT,
-            domainHint: TEST_CONFIG.DOMAIN_HINT,
-            claims: TEST_CONFIG.CLAIMS,
-            nonce: TEST_CONFIG.NONCE,
-            correlationId: RANDOM_TEST_GUID,
-            authenticationScheme: AuthenticationScheme.POP,
-            platformBroker: true,
-        };
-        const loginUrl = await client.getAuthCodeUrl(authCodeUrlRequest);
-        expect(
-            loginUrl.includes(
-                `${AADServerParamKeys.NATIVE_BROKER}=${encodeURIComponent("1")}`
-            )
-        ).toBe(true);
-        expect(
-            loginUrl.includes(
-                `${AADServerParamKeys.REQ_CNF}=${encodeURIComponent(
-                    TEST_POP_VALUES.ENCODED_REQ_CNF
-                )}`
-            )
-        ).toBe(true);
-    });
-
-    describe("handleFragmentResponse()", () => {
-        it("returns valid server code response", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-
-            const client: AuthorizationCodeClient = new AuthorizationCodeClient(
-                config
-            );
-            const authCodePayload = client.handleFragmentResponse(
-                {
-                    code: "thisIsATestCode",
-                    state: TEST_STATE_VALUES.ENCODED_LIB_STATE,
-                    client_info: TEST_DATA_CLIENT_INFO.TEST_RAW_CLIENT_INFO,
-                },
-                TEST_STATE_VALUES.ENCODED_LIB_STATE
-            );
-            expect(authCodePayload.code).toBe("thisIsATestCode");
-            expect(authCodePayload.state).toBe(
-                TEST_STATE_VALUES.ENCODED_LIB_STATE
-            );
-        });
-
-        it("throws server error when error is in hash", async () => {
-            jest.spyOn(
-                Authority.prototype,
-                <any>"getEndpointMetadataFromNetwork"
-            ).mockResolvedValue(DEFAULT_OPENID_CONFIG_RESPONSE.body);
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client: AuthorizationCodeClient = new AuthorizationCodeClient(
-                config
-            );
-            const cacheStorageMock =
-                config.storageInterface as MockStorageClass;
-
-            let error: AuthError | null = null;
-            try {
-                client.handleFragmentResponse(
-                    {
-                        error: "error_code",
-                        error_description: "msal error description",
-                        state: TEST_STATE_VALUES.ENCODED_LIB_STATE,
-                    },
-                    TEST_STATE_VALUES.ENCODED_LIB_STATE
-                );
-            } catch (e) {
-                error = e as AuthError;
-            }
-            expect(error).toBeInstanceOf(ServerError);
-            expect(error?.errorCode).toEqual("error_code");
-            expect(error?.errorMessage).toEqual("msal error description");
-            expect(cacheStorageMock.getKeys().length).toBe(1);
-            expect(cacheStorageMock.getAuthorityMetadataKeys().length).toBe(1);
         });
     });
 
@@ -3954,168 +2563,6 @@ describe("AuthorizationCodeClient unit tests", () => {
                 AADServerParamKeys.STATE
             }=test_state&testParam=test_val`;
             expect(logoutUri).toBe(testLogoutUriWithParams);
-        });
-    });
-
-    describe("createAuthCodeUrlQueryString tests", () => {
-        it("pick up default client_id", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                });
-
-            expect(queryString).toContain(
-                `client_id=${TEST_CONFIG.MSAL_CLIENT_ID}`
-            );
-        });
-
-        it("pick up extra query client_id param", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                    extraQueryParameters: {
-                        client_id: "child_client_id",
-                    },
-                });
-
-            expect(queryString).toContain(`client_id=child_client_id`);
-        });
-
-        it("pick up instance_aware config param when set to true", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            config.authOptions.instanceAware = true;
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                });
-
-            expect(queryString).toContain(`instance_aware=true`);
-        });
-
-        it("do not pick up instance_aware config param when set to false", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            config.authOptions.instanceAware = false;
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                });
-
-            expect(queryString.includes("instance_aware")).toBeFalsy();
-        });
-
-        it("pick up instance_aware EQ param when config is set to false", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            config.authOptions.instanceAware = false;
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                    extraQueryParameters: {
-                        instance_aware: "true",
-                    },
-                });
-
-            expect(queryString).toContain(`instance_aware=true`);
-        });
-
-        it("pick up instance_aware EQ param when config is set to true", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            config.authOptions.instanceAware = true;
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    prompt: PromptValue.LOGIN,
-                    redirectUri: "localhost",
-                    extraQueryParameters: {
-                        instance_aware: "false",
-                    },
-                });
-
-            expect(queryString).toContain(`instance_aware=false`);
-        });
-
-        it("pick up broker params", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    redirectUri: "localhost",
-                    embeddedClientId: "child_client_id_1",
-                });
-
-            expect(queryString).toContain(`client_id=child_client_id_1`);
-            expect(queryString).toContain(
-                `brk_client_id=${config.authOptions.clientId}`
-            );
-            expect(queryString).toContain(
-                `brk_redirect_uri=${encodeURIComponent("https://localhost")}`
-            );
-        });
-
-        it("broker params take precedence over extra query params", async () => {
-            const config: ClientConfiguration =
-                await ClientTestUtils.createTestClientConfiguration();
-            const client = new AuthorizationCodeClient(config);
-
-            const queryString =
-                // @ts-ignore
-                await client.createAuthCodeUrlQueryString({
-                    scopes: ["User.Read"],
-                    redirectUri: "localhost",
-                    embeddedClientId: "child_client_id_1",
-                    extraQueryParameters: {
-                        client_id: "child_client_id_2",
-                        brk_client_id: "broker_client_id_2",
-                        brk_redirect_uri: "broker_redirect_uri_2",
-                    },
-                });
-
-            expect(queryString).toContain(`client_id=child_client_id_1`);
-            expect(queryString).toContain(
-                `brk_client_id=${config.authOptions.clientId}`
-            );
-            expect(queryString).toContain(
-                `brk_redirect_uri=${encodeURIComponent("https://localhost")}`
-            );
         });
     });
