+++ lib/msal-node/src/client/ClientApplication.ts
@@ -57,6 +57,7 @@ import { version, name } from "../packageMetadata.js";
 import { UsernamePasswordRequest } from "../request/UsernamePasswordRequest.js";
 import { NodeAuthError } from "../error/NodeAuthError.js";
 import { UsernamePasswordClient } from "./UsernamePasswordClient.js";
+import { getAuthCodeRequestUrl } from "../protocol/Authorize.js";
 
 /**
  * Base abstract class for all ClientApplications - public and confidential
@@ -132,22 +133,18 @@ export abstract class ClientApplication {
             authenticationScheme: AuthenticationScheme.BEARER,
         };
 
-        const authClientConfig = await this.buildOauthClientConfiguration(
+        const discoveredAuthority = await this.createAuthority(
             validRequest.authority,
             validRequest.correlationId,
-            validRequest.redirectUri,
-            undefined,
             undefined,
             request.azureCloudOptions
         );
-        const authorizationCodeClient = new AuthorizationCodeClient(
-            authClientConfig
+        return getAuthCodeRequestUrl(
+            this.config,
+            discoveredAuthority,
+            validRequest,
+            this.logger
         );
-        this.logger.verbose(
-            "Auth code client created",
-            validRequest.correlationId
-        );
-        return authorizationCodeClient.getAuthCodeUrl(validRequest);
     }
 
     /**
@@ -180,13 +177,17 @@ export abstract class ClientApplication {
             validRequest.correlationId
         );
         try {
-            const authClientConfig = await this.buildOauthClientConfiguration(
+            const discoveredAuthority = await this.createAuthority(
                 validRequest.authority,
                 validRequest.correlationId,
-                validRequest.redirectUri,
-                serverTelemetryManager,
                 undefined,
                 request.azureCloudOptions
+            );
+            const authClientConfig = await this.buildOauthClientConfiguration(
+                discoveredAuthority,
+                validRequest.correlationId,
+                validRequest.redirectUri,
+                serverTelemetryManager
             );
             const authorizationCodeClient = new AuthorizationCodeClient(
                 authClientConfig
@@ -233,14 +234,18 @@ export abstract class ClientApplication {
             validRequest.correlationId
         );
         try {
+            const discoveredAuthority = await this.createAuthority(
+                validRequest.authority,
+                validRequest.correlationId,
+                undefined,
+                request.azureCloudOptions
+            );
             const refreshTokenClientConfig =
                 await this.buildOauthClientConfiguration(
-                    validRequest.authority,
+                    discoveredAuthority,
                     validRequest.correlationId,
                     validRequest.redirectUri || "",
-                    serverTelemetryManager,
-                    undefined,
-                    request.azureCloudOptions
+                    serverTelemetryManager
                 );
             const refreshTokenClient = new RefreshTokenClient(
                 refreshTokenClientConfig
@@ -283,14 +288,18 @@ export abstract class ClientApplication {
         );
 
         try {
+            const discoveredAuthority = await this.createAuthority(
+                validRequest.authority,
+                validRequest.correlationId,
+                undefined,
+                request.azureCloudOptions
+            );
             const clientConfiguration =
                 await this.buildOauthClientConfiguration(
-                    validRequest.authority,
+                    discoveredAuthority,
                     validRequest.correlationId,
                     validRequest.redirectUri || "",
-                    serverTelemetryManager,
-                    undefined,
-                    validRequest.azureCloudOptions
+                    serverTelemetryManager
                 );
             const silentFlowClient = new SilentFlowClient(clientConfiguration);
             this.logger.verbose(
@@ -391,14 +400,18 @@ export abstract class ClientApplication {
             validRequest.correlationId
         );
         try {
+            const discoveredAuthority = await this.createAuthority(
+                validRequest.authority,
+                validRequest.correlationId,
+                undefined,
+                request.azureCloudOptions
+            );
             const usernamePasswordClientConfig =
                 await this.buildOauthClientConfiguration(
-                    validRequest.authority,
+                    discoveredAuthority,
                     validRequest.correlationId,
                     "",
-                    serverTelemetryManager,
-                    undefined,
-                    request.azureCloudOptions
+                    serverTelemetryManager
                 );
             const usernamePasswordClient = new UsernamePasswordClient(
                 usernamePasswordClientConfig
@@ -465,29 +478,14 @@ export abstract class ClientApplication {
      * @param serverTelemetryManager - initializes servertelemetry if passed
      */
     protected async buildOauthClientConfiguration(
-        authority: string,
+        discoveredAuthority: Authority,
         requestCorrelationId: string,
         redirectUri: string,
-        serverTelemetryManager?: ServerTelemetryManager,
-        azureRegionConfiguration?: AzureRegionConfiguration,
-        azureCloudOptions?: AzureCloudOptions
+        serverTelemetryManager?: ServerTelemetryManager
     ): Promise<ClientConfiguration> {
         this.logger.verbose(
             "buildOauthClientConfiguration called",
             requestCorrelationId
-        );
-
-        // precedence - azureCloudInstance + tenant >> authority and request  >> config
-        const userAzureCloudOptions = azureCloudOptions
-            ? azureCloudOptions
-            : this.config.auth.azureCloudOptions;
-
-        // using null assertion operator as we ensure that all config values have default values in buildConfiguration()
-        const discoveredAuthority = await this.createAuthority(
-            authority,
-            requestCorrelationId,
-            azureRegionConfiguration,
-            userAzureCloudOptions
         );
 
         this.logger.info(
@@ -640,7 +638,7 @@ export abstract class ClientApplication {
      * object. If no authority set in application object, then default to common authority.
      * @param authorityString - authority from user configuration
      */
-    private async createAuthority(
+    protected async createAuthority(
         authorityString: string,
         requestCorrelationId: string,
         azureRegionConfiguration?: AzureRegionConfiguration,
@@ -651,7 +649,7 @@ export abstract class ClientApplication {
         // build authority string based on auth params - azureCloudInstance is prioritized if provided
         const authorityUrl = Authority.generateAuthority(
             authorityString,
-            azureCloudOptions
+            azureCloudOptions || this.config.auth.azureCloudOptions
         );
 
         const authorityOptions: AuthorityOptions = {
