+++ lib/msal-node/src/client/PublicClientApplication.ts
@@ -21,7 +21,7 @@ import {
     NativeSignOutRequest,
     AccountInfo,
     INativeBrokerPlugin,
-    ServerAuthorizationCodeResponse,
+    AuthorizeResponse,
     AADServerParamKeys,
     ServerTelemetryManager,
 } from "@azure/msal-common/node";
@@ -113,13 +113,17 @@ export class PublicClientApplication
             validRequest.correlationId
         );
         try {
-            const deviceCodeConfig = await this.buildOauthClientConfiguration(
+            const discoveredAuthority = await this.createAuthority(
                 validRequest.authority,
                 validRequest.correlationId,
-                "",
-                serverTelemetryManager,
                 undefined,
                 request.azureCloudOptions
+            );
+            const deviceCodeConfig = await this.buildOauthClientConfiguration(
+                discoveredAuthority,
+                validRequest.correlationId,
+                "",
+                serverTelemetryManager
             );
             const deviceCodeClient = new DeviceCodeClient(deviceCodeConfig);
             this.logger.verbose(
@@ -181,7 +185,7 @@ export class PublicClientApplication
         const loopbackClient: ILoopbackClient =
             customLoopbackClient || new LoopbackClient();
 
-        let authCodeResponse: ServerAuthorizationCodeResponse = {};
+        let authCodeResponse: AuthorizeResponse = {};
         let authCodeListenerError: AuthError | null = null;
         try {
             const authCodeListener = loopbackClient
