+++ lib/msal-node/test/client/ConfidentialClientApplication.spec.ts
@@ -360,7 +360,7 @@ describe("ConfidentialClientApplication", () => {
             };
 
             let acquireTokenByClientCredentialSpy: jest.SpyInstance;
-            let buildOauthClientConfigurationSpy: jest.SpyInstance;
+            let createAuthoritySpy: jest.SpyInstance;
             let sendPostRequestAsyncSpy: jest.SpyInstance;
             let client: ConfidentialClientApplication;
             let request: ClientCredentialRequest;
@@ -370,9 +370,9 @@ describe("ConfidentialClientApplication", () => {
                     <any>"acquireTokenByClientCredential"
                 );
 
-                buildOauthClientConfigurationSpy = jest.spyOn(
+                createAuthoritySpy = jest.spyOn(
                     ConfidentialClientApplication.prototype,
-                    <any>"buildOauthClientConfiguration"
+                    <any>"createAuthority"
                 );
 
                 sendPostRequestAsyncSpy = jest.spyOn(
@@ -404,10 +404,9 @@ describe("ConfidentialClientApplication", () => {
                 expect(acquireTokenByClientCredentialSpy).toHaveBeenCalledTimes(
                     1
                 );
-                expect(
-                    buildOauthClientConfigurationSpy.mock.lastCall[4]
-                        .azureRegion
-                ).toEqual(process.env[MSAL_FORCE_REGION]);
+                expect(createAuthoritySpy.mock.lastCall[2].azureRegion).toEqual(
+                    process.env[MSAL_FORCE_REGION]
+                );
 
                 checkRegion(
                     sendPostRequestAsyncSpy.mock.lastCall[0],
@@ -427,10 +426,9 @@ describe("ConfidentialClientApplication", () => {
                 expect(acquireTokenByClientCredentialSpy).toHaveBeenCalledTimes(
                     1
                 );
-                expect(
-                    buildOauthClientConfigurationSpy.mock.lastCall[4]
-                        .azureRegion
-                ).toEqual(region);
+                expect(createAuthoritySpy.mock.lastCall[2].azureRegion).toEqual(
+                    region
+                );
 
                 checkRegion(sendPostRequestAsyncSpy.mock.lastCall[0], region);
             });
@@ -446,8 +444,7 @@ describe("ConfidentialClientApplication", () => {
                     1
                 );
                 expect(
-                    buildOauthClientConfigurationSpy.mock.lastCall[4]
-                        .azureRegion
+                    createAuthoritySpy.mock.lastCall[2].azureRegion
                 ).toBeUndefined();
 
                 const endpoint: string =
